test(host): fold the stray createFakeHost suite into testing.test.ts
fake-host.test.ts and testing.test.ts both covered createFakeHost from
./testing, and no fake-host.ts has existed for some time. The setting-tab
assertions move in beside the vault and metadata ones as their own
concern-scoped describe.
